pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-ghost"
version = "1.0.0"
description = "Production-grade autonomous AI companion with emotional modeling and semantic memory"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Project Ghost Team"}
]
keywords = ["ai", "chatbot", "discord", "llm", "memory", "emotional-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "discord.py>=2.3.2",
    "aiohttp>=3.9.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "psutil>=5.9.8",
    "nvidia-ml-py>=12.535.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "tenacity>=8.2.3",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
ghost = "main:main"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]